Extract middleware shared examples
